New version with all files
New queries added so that when the create UPCID button is pressed the listbox on the left is populated with the UPCIDs
that are not in the selected table and the list box on the right is populated with the UPCIDs that are currently in the
selected table. Need to implement this into all the list boxes.