New upload with list box fixed
Query finds which UPCIDs columns are in the selected table already and store that as string on open of form selectUPCIDs.
User can now add UPCIDs from the left list box without any false records being added.

Now need to find a way to update the table with adding UPCIDs that are not in the table but are in the list box. Maybe using the string of UPCIDs